tools/requirements.txt
supabase
python-dotenv
pandas
openpyxl

